@@ -8,6 +8,8 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import com.project.shopapp.models.*;
 
+import java.util.List;
+
 public interface IProductService {
     Product createProduct(ProductDTO productDTO) throws Exception;
     Product getProductById(long id) throws Exception;
@@ -20,6 +22,5 @@ public interface IProductService {
             Long productId,
             ProductImageDTO productImageDTO) throws Exception;
 
-
+    List<Product> findProductsByIds(List<Long> productIds);
 }
-
