@@ -9,19 +9,33 @@ import {
 } from './components/detail-product/detail-product.component';
 import { OrderComponent } from './components/order/order.component';
 import { OrderDetailComponent } from './components/detail-order/order.detail.component';
-import { AuthGuardFn } from './components/guards/auth.guard';
+import { UserProfileComponent } from './components/user-profile/user.profile.component';
+import { AdminComponent } from './components/admin/admin.component';
+import { AuthGuardFn } from './guards/auth.guard';
+import { AdminGuardFn } from './guards/admin.guard';
+//import { OrderAdminComponent } from './components/admin/order/order.admin.component';
 
 const routes: Routes = [
   { path: '', component: HomeComponent },
-  { path: 'login', component: LoginComponent },
+  { path: 'login', component: LoginComponent },  
   { path: 'register', component: RegisterComponent },
-  { path: 'products/:id', component: DetailProductComponent },
+  { path: 'products/:id', component: DetailProductComponent },  
   { path: 'orders', component: OrderComponent,canActivate:[AuthGuardFn] },
+  { path: 'user-profile', component: UserProfileComponent, canActivate:[AuthGuardFn] },
   { path: 'orders/:id', component: OrderDetailComponent },
+  //Admin   
+  { 
+    path: 'admin', 
+    component: AdminComponent, 
+    canActivate:[AdminGuardFn] 
+  },      
 ];
 
 @NgModule({
-  imports: [RouterModule.forRoot(routes)],
+  imports: [
+    RouterModule.forRoot(routes),
+    CommonModule
+  ],
   exports: [RouterModule]
 })
 export class AppRoutingModule { }
