@@ -1,8 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 import { UserService } from '../../services/user.service';
 import { UserResponse } from 'src/app/responses/user/user.response';
-import { NgbPopoverConfig } from '@ng-bootstrap/ng-bootstrap';
 import { TokenService } from 'src/app/services/token.service';
+
+import { ActivatedRoute, Router } from '@angular/router';
 
 @Component({
   selector: 'app-header',
@@ -12,6 +13,18 @@ import { TokenService } from 'src/app/services/token.service';
 export class HeaderComponent implements OnInit{
   userResponse?:UserResponse | null;
   isPopoverOpen = false;
+  activeNavItem: number = 0;
+
+  constructor(
+    private userService: UserService,       
+    private tokenService: TokenService,    
+    private router: Router,
+  ) {
+    
+   }
+  ngOnInit() {
+    this.userResponse = this.userService.getUserResponseFromLocalStorage();    
+  }  
 
   togglePopover(event: Event): void {
     event.preventDefault();
@@ -20,7 +33,10 @@ export class HeaderComponent implements OnInit{
 
   handleItemClick(index: number): void {
     //alert(`Clicked on "${index}"`);
-    if(index === 2) {
+    if(index === 0) {
+      debugger
+      this.router.navigate(['/user-profile']);                      
+    } else if (index === 2) {
       this.userService.removeUserFromLocalStorage();
       this.tokenService.removeToken();
       this.userResponse = this.userService.getUserResponseFromLocalStorage();    
@@ -28,14 +44,9 @@ export class HeaderComponent implements OnInit{
     this.isPopoverOpen = false; // Close the popover after clicking an item    
   }
 
-  constructor(
-    private userService: UserService,   
-    private popoverConfig: NgbPopoverConfig,  
-    private tokenService: TokenService  
-  ) {
-    
-   }
-  ngOnInit() {
-    this.userResponse = this.userService.getUserResponseFromLocalStorage();    
+  
+  setActiveNavItem(index: number) {    
+    this.activeNavItem = index;
+    //alert(this.activeNavItem);
   }  
 }
