@@ -17,8 +17,19 @@ import { UserResponse } from '../../responses/user/user.response';
 export class LoginComponent implements OnInit{
   @ViewChild('loginForm') loginForm!: NgForm;
 
+  /*
+  //Login user
   phoneNumber: string = '33445566';
-  password: string = '123456';
+  password: string = '123456789';
+
+  //Login admin
+  phoneNumber: string = '11223344';
+  password: string = '11223344';
+
+  */
+  phoneNumber: string = '33445566';
+  password: string = '123456789';
+  showPassword: boolean = false;
 
   roles: Role[] = []; // Mảng roles
   rememberMe: boolean = true;
@@ -86,7 +97,12 @@ export class LoginComponent implements OnInit{
                 date_of_birth: new Date(response.date_of_birth),
               };    
               this.userService.saveUserResponseToLocalStorage(this.userResponse); 
-              this.router.navigate(['/']);                      
+              if(this.userResponse?.role.name == 'admin') {
+                this.router.navigate(['/admin']);    
+              } else if(this.userResponse?.role.name == 'user') {
+                this.router.navigate(['/']);                      
+              }
+              
             },
             complete: () => {
               debugger;
@@ -106,5 +122,8 @@ export class LoginComponent implements OnInit{
         alert(error.error.message);
       }
     });
+  }
+  togglePassword() {
+    this.showPassword = !this.showPassword;
   }
 }
